implement IdentityMessage; send MSG_IDENT on accept

+++ src/flags.hpp
@@ -30,8 +30,8 @@ set_listen_string(std::string a, std::string b, void* d) {
 
 void
 set_id(std::string a, std::string b, void* d) {
-	auto id = reinterpret_cast<int*>(d);
-	*id = atoi(b.c_str());
+	auto id = reinterpret_cast<uint64_t*>(d);
+	*id = (uint64_t)atoi(b.c_str());
 }
 
 void
